WIP - Experimental - Refactored search - Created a search subpackage... - ...gathering Search and SearchMeta classes - Promote common behaviors in SearchConfig and SearchBase classes
