+++ monggregate/stages/search/__init__.py
@@ -1,0 +1,8 @@
+"""Search stages subpackage.
+
+Contains search and search_meta modules.
+
+"""
+
+from monggregate.stages.search.search import Search
+from monggregate.stages.search.search_meta import SearchMeta
