+++ monggregate/stages/search/base.py
@@ -1,72 +1,4 @@
-"""Module definining an interface to MongoDB $search stage operation in aggregation pipeline.
-
-Online MongoDB documentation:
---------------------------------------------------------------------------------------------------
-
-Last Updated (in this package) : 25/04/2023
-Source : https://www.mongodb.com/docs/atlas/atlas-search/query-syntax/#mongodb-pipeline-pipe.-search
-
-# Definition
-#---------------------------
-The $search stage performs a full-text search on the specified field or fields which must be covered by an Atlas Search index.
-
-$search
-A $search pipeline stage has the following prototype form:
-
-    >>> {
-            $search: {
-                "index": "<index-name>",
-                "<operator-name>"|"<collector-name>": {
-                <operator-specification>|<collector-specification>
-                },
-                "highlight": {
-                <highlight-options>
-                },
-                "count": {
-                <count-options>
-                },
-                "returnStoredSource": true | false
-            }
-        }
-
-# Fields
-#---------------------------
-
-The $search stage takes a document with the following fields
-
-Field                       Type       Necessity       Description
-
-<collector-name>            document   Conditional     Name of the collector to use with the query. 
-                                                       You can provide a document that contains the collector-specific options as the value for this field. 
-                                                       Either this or <operator-name> is required.
-count                       document   Optional        Document that specifies the count options for retrieving a count of the results. 
-                                                       To learn more, see Count Atlas Search Results.
-highlight                   document   Optional        Document that specifies the highlight options for displaying search terms in their original context.
-index                       string     Required        Name of the Atlas Search index to use. If omitted, defaults to default
-<operator-name>             document   Conditional     Name of the operator to search with. 
-                                                       You can provide a document that contains the operator-specific options as the value for this field. 
-                                                       Either this or <collector-name> is required.
-returnStoredSource          boolean    Optional        Flag that specifies whether to perform a full document lookup on the backend database or return only stored source fields directly from Atlas Search. 
-                                                       If omitted, defaults to false. To learn more, see Return Stored Source pyd.Fields.
-
-# Behavior
-#---------------------------
-$search must be the first stage of any pipeline it appears in. 
-$search cannot be used in:
-
-    * a view definition
-
-    * a $facet pipeline stage
-
-# Aggregation Variable
-#---------------------------
-$search returns only the results of your query. The metadata results of your 
-$search query are saved in the $$SEARCH_META aggregation variable. You can use the $$SEARCH_META variable to view the metadata results for your 
-$search query. The $$SEARCH_META aggregation variable can be used anywhere after a 
-$search stage in any pipeline, but it can't be used after the $lookup or $unionWith stage in any pipeline. 
-The $$SEARCH_META aggregation variable can't be used in any subsequent stage after a $searchMeta stage.
-                                                       
-"""
+"""xxx"""
 
 from datetime import datetime
 from typing import Any, Callable, Literal
@@ -75,7 +7,7 @@ try:
 except ImportError:
     from typing_extensions import Self
     
-from monggregate.base import pyd
+from monggregate.base import pyd, BaseModel
 from monggregate.stages.stage import Stage
 from monggregate.search.collectors import Facet, Facets
 from monggregate.search.operators import(
@@ -108,79 +40,28 @@ OperatorLiteral = Literal[
     "wildcard"
 ]
 
-# Classes
-# -----------------------------------------------------
-class SearchBase(Stage):
+
+class SearchConfig(BaseModel):
     """Internals"""
 
     index : str = "default"
     count : dict|None
     highlight : dict|None
-    return_stored_source : bool = pyd.Field(False, alias="returnStoredSource")
-    score_details : bool = pyd.Field(False, alias="scoreDetails")
+    return_stored_source : bool = False
+    score_details : bool  = False
 
     @property
-    def statement(self) -> dict[str, dict]:
-    
-        config = {
-                "index":self.index,
-                "highlight":self.highlight,
-                "count":self.count,
-                "returnStoredSource":self.return_stored_source,
-                "scoreDetails":self.score_details
-            }
-        
-     
-        _statement = {
-            "$search":config
-        }
-     
-        return self.resolve(_statement)
+    def statement():
+        """Returns the statement of the stage"""
 
+        raise NotImplementedError("statement method must be implemented in subclasses")
 
-class Search(SearchBase):
-    """"
-    Creates a $search statement in an aggregation pipeline
+class SearchBase(Stage, SearchConfig):
+    """xxxx"""
 
-    Descrtiption
-    -----------------------
-    The $search stage performs a full-text search on the specified field or fields 
-    which must be covered by an Atlas Search index.
-
-    Attributes:
-    -----------------------
-        - index, str : name of the Atlas Search index to use. Defaults to default.
-
-        - count, dict|None : Document that specifies the count options for retrieving a count
-                             of the results. 
-
-        - highlight, dict|None : Document that specifies the highlight options for displaying
-                                 search terms in their original context.
-
-        - return_stored_source, bool : Flag that specifies whether to perform a full document lookup
-                                       on the backend database (mongod) or return only stored source fields
-                                       directly from Atlas Search. Defaults to false.
-
-        - score_details, bool : Flag that specifies whether to retrieve a detailed breakdown of
-                                the score for the documents in the results. Defaults to false
-                                To view the details, you must use the $meta expression in the
-                                $project stage.
-
-        - <operator-name>, dict|None : Name of the operator to search with. You can provide a document
-                                  that contains the operator-specific options as the value for this field
-                                  Either this or <collector-name> is required.
-
-        - <collector-name>, dict|None : Name of the collector to use with the query. You can provide
-                                        a document that contains the collector-specific options as the value
-                                        for this field. Either this or <operator-name> is required.
-
-    """
-
-    
     collector : Facet|None
     operator : AnyOperator|None
-    minimum_should_match : int = 0
-
+    minimum_should_match : int = 0 # specific to Compound
 
     @pyd.root_validator(pre=True)
     @classmethod
@@ -210,28 +91,13 @@ class Search(SearchBase):
         return value
     
     @property
-    def statement(self) -> dict[str, dict]:
-    
-        config = {
-                "index":self.index,
-                "highlight":self.highlight,
-                "count":self.count,
-                "returnStoredSource":self.return_stored_source,
-                "scoreDetails":self.score_details
-            }
-        
-        method = self.collector or self.operator
+    def statement():
+        """Returns the statement of the stage"""
 
-        config.update(method.statement)
-
-        _statement = {
-            "$search":config
-        }
-     
-        return self.resolve(_statement)
+        raise NotImplementedError("statement method must be implemented in subclasses")
     
 
-    #---------------------------------------------------------
+     #---------------------------------------------------------
     # Constructors
     #---------------------------------------------------------
     @classmethod
@@ -275,7 +141,7 @@ class Search(SearchBase):
 
         """
 
-        base_params = SearchBase(**kwargs).dict()
+        base_params = SearchConfig(**kwargs).dict()
         cls.__reduce_kwargs(kwargs)
 
         autocomplete_statement = Autocomplete(
@@ -302,7 +168,7 @@ class Search(SearchBase):
         
     )->Self:
 
-        base_params = SearchBase(**kwargs).dict()
+        base_params = SearchConfig(**kwargs).dict()
         cls.__reduce_kwargs(kwargs)
 
         compound_statement = Compound(
@@ -335,7 +201,7 @@ class Search(SearchBase):
         
         """
 
-        base_params = SearchBase(**kwargs).dict()
+        base_params = SearchConfig(**kwargs).dict()
         equals_statement = Equals(
             path=path,
             value=value,
@@ -357,7 +223,7 @@ class Search(SearchBase):
         
         """
 
-        base_params = SearchBase(**kwargs).dict()
+        base_params = SearchConfig(**kwargs).dict()
         exists_statement = Exists(path=path)
 
         return cls(**base_params, operator=exists_statement)
@@ -372,7 +238,7 @@ class Search(SearchBase):
 
         """
         
-        base_params = SearchBase(**kwargs).dict()
+        base_params = SearchConfig(**kwargs).dict()
         cls.__reduce_kwargs(kwargs)
         
         operator = kwargs.pop("operator", None)
@@ -393,7 +259,7 @@ class Search(SearchBase):
 
         """
         
-        base_params = SearchBase(**kwargs).dict()
+        base_params = SearchConfig(**kwargs).dict()
         more_like_this_stasement = MoreLikeThis(like=like)
 
         return cls(**base_params, operator=more_like_this_stasement)
@@ -421,7 +287,7 @@ class Search(SearchBase):
         
         """
 
-        base_params = SearchBase(**kwargs).dict()
+        base_params = SearchConfig(**kwargs).dict()
         range_statement = Range(
             path=path,
             gt=gt,
@@ -451,7 +317,7 @@ class Search(SearchBase):
         
         """
 
-        base_params = SearchBase(**kwargs).dict()
+        base_params = SearchConfig(**kwargs).dict()
         regex_statement = Regex(
             query=query,
             path=path,
@@ -481,7 +347,7 @@ class Search(SearchBase):
         
         """
 
-        base_params = SearchBase(**kwargs).dict()
+        base_params = SearchConfig(**kwargs).dict()
         cls.__reduce_kwargs(kwargs)
 
         text_statement = Text(
@@ -512,7 +378,7 @@ class Search(SearchBase):
         
         """
 
-        base_params = SearchBase(**kwargs).dict()
+        base_params = SearchConfig(**kwargs).dict()
         cls.__reduce_kwargs(kwargs)
         
         wilcard_statement = Wildcard(
@@ -752,24 +618,3 @@ class Search(SearchBase):
         kwargs.pop("highlight", None)
         kwargs.pop("return_stored_source", None)
         kwargs.pop("score_details", None)
-
-
-
-if __name__ == "__main__":
-    search = Search(index="fruits")
-    search.text(
-    "should", 
-    query="apple", 
-    path="type"
-    ).compound(
-        "should"
-    ).text(
-        "must",
-        query="organic",
-        path="category"
-   ).equals(
-        "must",
-        value=True,
-        path="in_stock"
-    )
-    print(search)
