+++ monggregate/stages/search/search_meta.py
@@ -63,10 +63,10 @@ facet           The result to a facet query is a mapping of the defined facet na
 
 """
 
-from monggregate.stages.search import Search
+from monggregate.stages.search.base import SearchConfig, SearchBase
 
 
-class SearchMeta(Search):
+class SearchMeta(SearchBase):
     """
     Creates a $searchMeta statement in an aggregation pipeline
